[Html] 1. header, footer 추가

+++ src/main/java/com/jang/board/service/PostServiceImpl.java
@@ -16,10 +16,12 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpSession;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -98,23 +100,24 @@ public class PostServiceImpl implements PostService {
         Post post = findPost(id);
 
         List<Photo> photos = new ArrayList<>(post.getPhotos());
-        for(Photo photo : photos) {
-            photoRepository.delete(photo);
-        }
+        if(!CollectionUtils.isEmpty(originalFilenames) && !originalFilenames.get(0).equals("")) {
+            for (Photo photo : photos) {
+                photoRepository.delete(photo);
+            }
 
-        photos = new ArrayList<>();
-        for(int i = 0; i < originalFilenames.size(); i++) {
-            Photo photo = Photo.createPhoto(post.getMember(), originalFilenames.get(i), storeFilenames.get(i));
-            photos.add(photo);
-            photo.changePost(post);
-            photoRepository.save(photo);
+            photos = new ArrayList<>();
+            for (int i = 0; i < originalFilenames.size(); i++) {
+                Photo photo = Photo.createPhoto(post.getMember(), originalFilenames.get(i), storeFilenames.get(i));
+                photos.add(photo);
+                photo.changePost(post);
+                photoRepository.save(photo);
+            }
+
+            fileStore.deleteFile(photos.stream()
+                    .map(Photo::getStoreFileName)
+                    .collect(Collectors.toList()));
         }
 
         post.changePost(title, content, photos);
-
-        fileStore.deleteFile(photos.stream()
-                .map(Photo::getStoreFileName)
-                .collect(Collectors.toList()));
-
     }
 }
